+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/notifications/Notification.kt
@@ -50,6 +50,7 @@ data class Notification(
         )
 
         fun noCommitMessage(): Notification = Notification(message = message("notifications.no-commit-message"))
+        fun unableToSaveToken(): Notification = Notification(message = message("notifications.unable-to-save-token"))
 
         fun unableToSaveToken() = Notification(message = message("notifications.unable-to-save-token"))