+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/AppSettings.kt
@@ -1,7 +1,10 @@
 package com.github.blarc.ai.commits.intellij.plugin.settings
 
+import com.aallam.openai.client.OpenAIConfig
+import com.aallam.openai.client.ProxyConfig
 import com.github.blarc.ai.commits.intellij.plugin.notifications.Notification
 import com.github.blarc.ai.commits.intellij.plugin.notifications.sendNotification
+import com.github.blarc.ai.commits.intellij.plugin.settings.prompt.Prompt
 import com.intellij.credentialStore.CredentialAttributes
 import com.intellij.credentialStore.Credentials
 import com.intellij.ide.passwordSafe.PasswordSafe
@@ -10,7 +13,10 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.PersistentStateComponent
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
+import com.intellij.util.xmlb.Converter
 import com.intellij.util.xmlb.XmlSerializerUtil
+import com.intellij.util.xmlb.annotations.OptionTag
+import java.util.*
 
 @State(
     name = AppSettings.SERVICE_NAME,
@@ -33,6 +39,10 @@ class AppSettings : PersistentStateComponent<AppSettings> {
 
     var requestSupport = true
     var lastVersion: String? = null
+    var proxyUrl: String? = null
+
+    var prompts: MutableMap<String, Prompt> = initPrompts()
+    var currentPrompt: Prompt = prompts["basic"]!!
 
     companion object {
         const val SERVICE_NAME = "com.github.blarc.ai.commits.intellij.plugin.settings.AppSettings"
@@ -57,6 +67,11 @@ class AppSettings : PersistentStateComponent<AppSettings> {
         }
     }
 
+    fun getOpenAIConfig(): OpenAIConfig {
+        val token = getOpenAIToken() ?: throw Exception("OpenAI Token is not set.")
+        return OpenAIConfig(token, proxy = proxyUrl?.takeIf { it.isNotBlank() }?.let { ProxyConfig.Http(it) })
+    }
+
     fun getOpenAIToken(): String? {
         val credentialAttributes = getCredentialAttributes(openAITokenTitle)
         val credentials: Credentials = PasswordSafe.instance.get(credentialAttributes) ?: return null
@@ -65,10 +80,10 @@ class AppSettings : PersistentStateComponent<AppSettings> {
 
     private fun getCredentialAttributes(title: String): CredentialAttributes {
         return CredentialAttributes(
-            title,
-            null,
-            this.javaClass,
-            false
+                title,
+                null,
+                this.javaClass,
+                false
         )
     }
 
@@ -80,9 +95,56 @@ class AppSettings : PersistentStateComponent<AppSettings> {
 
     fun recordHit() {
         hits++
-        return
-        /* if (requestSupport && (hits == 50 || hits % 100 == 0)) {
+        if (requestSupport && (hits == 50 || hits % 100 == 0)) {
             sendNotification(Notification.star())
-        } */
+        }
+    }
+
+    private fun initPrompts() = mutableMapOf(
+            // Generate UUIDs for game objects in Mine.py and call the function in start_game().
+            "basic" to Prompt("Basic",
+                    "Basic prompt that generates a decent commit message.",
+                    "Write an insightful but concise Git commit message in a complete sentence in present tense for the " +
+                            "following diff without prefacing it with anything, the response must be in the language {locale} and must " +
+                            "NOT be longer than 74 characters. The sent text will be the differences between files, where deleted lines" +
+                            " are prefixed with a single minus sign and added lines are prefixed with a single plus sign.\n" +
+                            "{diff}",
+                    false),
+            // feat: generate unique UUIDs for game objects on Mine game start
+            "conventional" to Prompt("Conventional",
+                    "Prompt for commit message in the conventional commit convention.",
+                    "Write a clean and comprehensive commit message in the conventional commit convention. " +
+                            "I'll send you an output of 'git diff --staged' command, and you convert " +
+                            "it into a commit message. " +
+                            "Do NOT preface the commit with anything. " +
+                            "Do NOT add any descriptions to the commit, only commit message. " +
+                            "Use the present tense. " +
+                            "Lines must not be longer than 74 characters. " +
+                            "Use {locale} language to answer.\n" +
+                            "{diff}",
+                    false),
+            // ✨ feat(mine): Generate objects UUIDs and start team timers on game start
+            "emoji" to Prompt("Emoji",
+                    "Prompt for commit message in the conventional commit convention with GitMoji convention.",
+                    "Write a clean and comprehensive commit message in the conventional commit convention. " +
+                            "I'll send you an output of 'git diff --staged' command, and you convert " +
+                            "it into a commit message. " +
+                            "Use GitMoji convention to preface the commit. " +
+                            "Do NOT add any descriptions to the commit, only commit message. " +
+                            "Use the present tense. " +
+                            "Lines must not be longer than 74 characters. " +
+                            "Use {locale} language to answer.\n" +
+                            "{diff}",
+                    false)
+    )
+
+    class LocaleConverter : Converter<Locale>() {
+        override fun toString(value: Locale): String? {
+            return value.toLanguageTag()
+        }
+
+        override fun fromString(value: String): Locale? {
+            return Locale.forLanguageTag(value)
+        }
     }
 }